# Feedback/FeedbackEventService/swagger.yaml
swagger: '2.0'
info:
  description: "Events & Reports"
  version: '1.0'
  title: Events
host: "local"
basePath: "/events"
schemes: 
  - http
paths:
  /dashboard:
    get:
      tags:
        - Dashboard
      summary: Get Dashboard
      description: Get the Details of All Events
      operationId: getDashboard
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: array
            $ref: '#/definitions/Event'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false 
...
  '/event':    
    post:
      tags:
        - Event
      summary: Add Event
      description: Add the Event Summary to Database
      operationId: addEvent
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: event
          required: true
          schema:
            $ref: '#/definitions/Event'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Event'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    get:
      tags:
        - Event
      summary: Get Event 
      description: Get the Details of All Events
      operationId: getAllEvent
      consumes:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Event'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  '/event/{eventId}':
    get:
      tags:
        - Event
      summary: Get Event by Id
      description: Provide an Id to look up for specific Event
      operationId: getEventById
      produces:
        - application/json
      parameters:
        - name: eventId
          in: path
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Event'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
...
definitions:
  Event:
    type: object
    properties:
      event_id:
        type: string
      month:
        type: string
      base_location:
        type: string
      beneficiary_name:
        type: string
      venture_address:
        type: string
      council_name:
        type: string
      project:
        type: string  
      category:
        type: string
      event_name:
         type: string
      event_description:
         type: string
      event_date:
        type: string
        format: date
      total_no_of_volunteers:
        type: integer
        format: int32
      total_volunteer_hours:
        type: integer
        format: int32
      total_travel_hours:
        type: integer
        format: int32
      overall_volunteer_hours:
        type: integer
        format: int32
      lives_impacted:
        type: integer
      activity_type:
        type: integer
      status:
        type: string
      poc_id:
        type: integer
        format: int32
      poc_name:
        type: string
      poc_contact:
        type: integer
        format: int64
    title: Event
...